[xbase] expanded operators for all combinations of primitive types https://bugs.eclipse.org/bugs/show_bug.cgi?id=367207 https://bugs.eclipse.org/bugs/show_bug.cgi?id=367205

+++ plugins/org.eclipse.xtext.xbase.junit/src/org/eclipse/xtext/xbase/junit/evaluation/AbstractXbaseEvaluationTest.java
@@ -355,7 +355,7 @@ public abstract class AbstractXbaseEvaluationTest extends TestCase {
 	}
 
 	@Test public void testPowerOnIntegers() throws Exception {
-		assertEvaluatesTo(new Integer(8), "2**3");
+		assertEvaluatesTo(new Double(8), "2**3");
 	}
 
 	@Test public void testLessThanOnIntegers_01() throws Exception {
